correção de repositorio, integração de home

+++ lib/app/shared/models/enterprise_model.dart
@@ -13,7 +13,7 @@ class EnterpriseModel {
   String? photo;
   int? value;
   int? shares;
-  int? sharePrice;
+  double? sharePrice;
   int? ownShares;
   String? city;
   String? country;
